refactor(controllers): consolidate authentication controllers and update imports
- Move web/auth controllers to top-level auth directory
- Remove common/security/keys.controller.js as it's no longer needed
- Update relative import paths to reflect new directory structure
- Fix SMSService import path for constants.util
- Adjust crud template import path for response helper
